allow numpy to use float64/float32

+++ src/pywavelet/types/wavelet_bins.py
@@ -1,6 +1,6 @@
 from typing import Tuple, Union
 
-from ..backend import xp
+from ..backend import xp, float_dtype, complex_dtype
 from .frequencyseries import FrequencySeries
 from .timeseries import TimeSeries
 
@@ -35,13 +35,7 @@ def _get_bins(
 ) -> Tuple[xp.ndarray, xp.ndarray]:
     T = data.duration
     t_bins, f_bins = compute_bins(Nf, Nt, T)
-
-    # N = len(data)
-    # fs = N / T
-    # assert delta_f == fmax / Nf, f"delta_f={delta_f} != fmax/Nf={fmax/Nf}"
-
     t_bins += data.t0
-
     return t_bins, f_bins
 
 
@@ -51,6 +45,6 @@ def compute_bins(Nf: int, Nt: int, T: float) -> Tuple[xp.ndarray, xp.ndarray]:
     """
     delta_T = T / Nt
     delta_F = 1 / (2 * delta_T)
-    t_bins = xp.arange(0, Nt) * delta_T
-    f_bins = xp.arange(0, Nf) * delta_F
+    t_bins = xp.arange(0, Nt, dtype=float_dtype) * delta_T
+    f_bins = xp.arange(0, Nf, dtype=float_dtype) * delta_F
     return t_bins, f_bins
